features added/bug fixes
added button to allow the user to pause the game and use a pause menu including buttons for resuming, saving loading, and exit to menu. Note: all but save load buttons are useable.
Bug Fixes: bug fixed with issues with enemy pathfinding, issue was relating to when player was loaded into a scene the ai chase script of the enemies could not find the player has not completely loaded in, so let ai wait a little until finding the player, also increased mass of enemies to reduce sliding when colliding with player.